Add expert view toggle (issue #50)
Expert view can be toggled in the context menu and the setting is
persisted on a per project basis. Open editors/views that are related
with expert view are currently still left untouched when the expert mode
is deactivated. The new feature only affects the display of the
Hierarchy node currently.

Change-Id: I8338d160cee18c5d6ef14ab916dd067dee41239f
Signed-off-by: Denis Knoepfle <denis.knoepfle@sap.com>