Replace used combo cell editor by custom variant
Now using the native combo box instead of the CCombo to get rid of
unnecessary white space in the item list of the combo box. Also clicking
on an item with the mouse automatically deactivates the editing properly
now (so that the dirty state of the editor page is shown correctly,
earlier it was necessary to make the control lose focus first or hit
enter to apply the value).

Change-Id: I2231fa7fa563891e8cdfc9a7ed1197b53caf4506
Signed-off-by: Denis Knoepfle <denis.knoepfle@sap.com>