Separate UI physically from service layer (#118)
The UI transmits now a JobDescription when a diagnosis should be
performed, so the server creates necessary configuration files for DS to
run (identical to those locally configured on the client-side before)
using this data. The results are only saved on the server side now. The
configuration files and results are saved on a per job-id basis, so that
(not implemented yet) the client can request the results later via DS
service using the job id that was used to trigger the diagnosis run.

Change-Id: I990cc93a2851726dead0d33e2f1a935ddd5c46ad
Signed-off-by: Denis Knoepfle <denis.knoepfle@sap.com>