Add service operation to retrieve results (#116)
Currently only the main data can be fetched, but a download option to
get all the raw data will be offered later. The results are not fetched
within the UI yet.

The SpotterResult's message field had to be renamed to comply with JSON.

Added the report text as a string inside ResultsContainer to have all
data in one spot. However, the report text file is written normally like
before. Changes in Spotter were only to set the report in the container
object.

Change-Id: I13b546924b0c0b7c832ee20f6c59f3f5f9244a50
Signed-off-by: Denis Knoepfle <denis.knoepfle@sap.com>