Fix small bug with expert view related to caching
When a project was set to expert view, deleted and another project with
the same name created, then the project's SpotterProjectParent object
remained cached in the content provider and thus had the unexpected
'Hierarchy' child. Now with the deletion of a project the corresponding
node gets removed from the cache properly.

Change-Id: I39e1a6aeaac43a97d7061da2f01092c675bfd08d
Signed-off-by: Denis Knoepfle <denis.knoepfle@sap.com>