Enhance fetching of results and other minor improvements
Now resources like images etc. will also be fetched, raw data is
ignored.

Improved the loading process of result items in the navigator. The
loading of the items will now be deferred and performed in the
background, so that the user has a quicker feedback when expanding the
project node for the first time.

When run results nodes are already displaying but their data has been
deleted for some reason, it will be automatically refetched before
trying to access the files. This only counts for the whole folder,
single files won't be checked.

Change-Id: I8d300c0e30f1bde3c4bb68e81c6ac667e8fcde04
Signed-off-by: Denis Knoepfle <denis.knoepfle@sap.com>