Improve Active Run View and fix some bugs
The Active Run View shows the hierarchy and the progress of each node
now. Currently child nodes that are skipped in the diagnosis are not
treated specially yet.

Fixed NullPointerException in RunExtensionsImageProvider (the spotter
progress was never set).

Fixed SWT invalid thread access violations and SWT exceptions if widgets
were disposed already in Active Run View.

Change-Id: I23125d3706b5b106e60451a5831c8ae0ae77480e
Signed-off-by: Denis Knoepfle <denis.knoepfle@sap.com>