Add erroneous run nodes and change fetch logic
Now erroneous diagnosis runs are also displayed in the project
navigator. Opening such an item shows the corresponding error message of
the failed diagnosis.

Changed the fetch logic a little:
If an empty (0 bytes) archive is received, it will be assumed that the
corresponding data is missing on the server, so the job id will be
deleted on the client side for that run.

Change-Id: If34500d8dc88b464c33767cd0f4273e435ba4824
Signed-off-by: Denis Knoepfle <denis.knoepfle@sap.com>